🔥 Remove browser env info
This information is not always present and can cause type errors when attempting to concatenate into a string. The only SDKs to capture this information is this SDK and the Java SDK. Since no other SDK captures this info and it is not needed in support, we can remove it to prevent type errors.