🔥 Remove browser env info (#8)
* 🔥 Remove browser env info

This information is not always present and can cause type errors when attempting to concatenate into a string. The only SDKs to capture this information is this SDK and the Java SDK. Since no other SDK captures this info and it is not needed in support, we can remove it to prevent type errors.

* 🔥 Remove unused arg

* 📦 Update chromedriver

* 👷 Remove python2 job

This job's image has a version of chrome that is different than the python3 image which causes one
of them to fail. Since python2 is EOL, we can safely remove this job and allow the python3 job to be
the only CI job.